Structured name mangling (#1372)
* Structured name mangling

Previously we used to mangle names early in the
pipeline using a lossy conversion to strings that
couldn't be reversed.

This pr introduces a few things to improve it:

1. Introduce new name mangling scheme with a small spec.

2. Global names now incorporate signatures, an additional
   data structure that lets one programmatically see additional
   information about the name.

3. Adds a compelete unmangler to make sure that we can
   go back to the unmangled representation from the mangled
   one.

* Make sure that #900 is fixed

Fixes #900

* Add support for nullability to NIR

* Address pull request feedback

* Fix global value numbering

* Fix file name typo

+++ nir/src/main/scala/scala/scalanative/nir/Fresh.scala
@@ -3,7 +3,7 @@ package nir
 
 import java.util.concurrent.atomic.AtomicInteger
 
-final class Fresh private (private var start: Int) {
+final class Fresh private (private var start: Long) {
   def apply(): Local = {
     start += 1
     val value = start
@@ -12,11 +12,11 @@ final class Fresh private (private var start: Int) {
 }
 
 object Fresh {
-  def apply(start: Int = 0): Fresh =
+  def apply(start: Long = 0L): Fresh =
     new Fresh(start)
 
   def apply(insts: Seq[Inst]): Fresh = {
-    var max = -1
+    var max = -1L
     insts.foreach {
       case Inst.Let(local, _, _) =>
         max = Math.max(max, local.id)
